Fix typing of MESS
Introduces the MessType and exposes ArrayValue in order to make
typing of MESS possible.

Before, there was a problem where the type `ReadArrayValue` was not
correct when it came to MESS. The problem was that read did not
return a value of type `MESS` but rather it returned the value `MESS`.
In order to fix this, we introduce the `MessType` so that `MESS`
has the type `MessType`.

Except for typing, this should be backwards compatible when
it comes to executable code, but id does break typing.